Move node config into lib/config
