+++ src/lib/config/node_config.ts
